refactor and style tags
